ADAPTOR/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
paho-mqtt==1.6.1
requests==2.31.0
urllib3==2.0.4
